refactor: share train+heldout orchestration; bring run_hpc.py to parity
The Beaker entry point run_hpc.py previously only did instantiate(model)+fit(),
so it silently skipped the held-out work the new configs request
(auto_heldout_finetune for disRNN/GRU, heldout_refit for baseline_rl, post-training
held-out eval, multisubject W&B naming) — all of which lived only in run_capsule.py.

Extract that body verbatim into code/training_runner.run_training(hydra_config,
wandb_run); both entry points now keep only their own config-loading and call it:
- run_capsule.py (Code Ocean): config from /data/jobs.
- run_hpc.py (Beaker/HPC): dispatcher config via relative path; now also applies
  penalty multipliers + dynamic run-name components before W&B, matching run_capsule.

Verified: all three files compile; full import chain imports (training_runner,
run_capsule, run_hpc); real run_hpc.py composes data=mice_snapshot x
{disrnn,gru,baseline_rl} via its relative-path contract (--cfg job, exit 0) with
auto_heldout_finetune / heldout_refit present in the resolved config.
